# optimize `permute!` for certain data types